refactor: extract interaction_types, chat_turn_heuristics, stop_hooks_yaml to astra-turn-core (Phase 23)
Phase 23 extracts shared interaction types and two dependent modules:

- interaction_types.rs (NEW): TurnInteractionMode, TurnInteractionPolicy,
  ASK_USER_TOOL_NAME — extracted from agentic_loop_host to unblock deps
- chat_turn_heuristics.rs (792 LOC): Factual-query detection, session error
  classification, task execution profiling
- stop_hooks_yaml.rs (708 LOC): YAML-based stop hook configuration parsing

agentic_loop_host now re-exports interaction types from turn-core.

Turn-core: 1817 tests passing
Runtime: 2968 tests + 5 pre-existing remote skill failures
