+++ rust/crates/astra-turn-core/src/lib.rs
@@ -20,6 +20,7 @@ pub mod cache_diagnostics;
 pub mod chat_history_openai;
 pub mod chat_turn_api_error;
 pub mod chat_turn_explain_wire;
+pub mod chat_turn_heuristics;
 pub mod confidence_contract;
 pub mod context_assembly_trace;
 pub mod edge_executor_id;
@@ -35,6 +36,7 @@ pub mod headless_tool_status_display;
 pub mod headless_tool_stderr_lines;
 pub mod history;
 pub mod hook_plans;
+pub mod interaction_types;
 pub mod interruption;
 pub mod microcompact;
 pub mod observer;
@@ -51,6 +53,7 @@ pub mod sse_blocks;
 pub mod sse_edge_stderr_lines;
 pub mod state;
 pub mod stop_hooks;
+pub mod stop_hooks_yaml;
 pub mod streaming_tool_exec;
 pub mod tail_persist;
 pub mod task;
